refactor: modify struct of pages directory
